Refactor and cleanup FileList
